A genetic algorithm for optimizing neural network parameters using Python
Artificial neural network is a supervised machine learning algorithm very popular in applications in various fields such as speech and image recognition, time series forecasting, machine translation software, among others. They are useful in researches by the ability to solve stochastic problems, which often allows for approximate solutions to extremely complex problems.
However, it is very difficult to define ideal network architecture because there are no clear rules for how many neurons in the intermediate layers or how many layers or how the connections between these neurons should be implemented. To solve this kind of problem, this article instructs how to use a Genetic Algorithm to automatically find good neural network architectures in Python.
First, you need to install the scikit-learn package. A Simple and efficient tool for data mining and data analysis.
For the training of the hybrid algorithm we will use a database of Iris flower classes (Setosa, Virginica and Versicolor).